feat: contract delegation tests
Adding support for "long" testing mode which basically runs costly
functions in a loop until they run out of gas and then return. The
goal is to be able to cause the client to run a lot of code for a very
low gas cost

Also made a minor change to the load test logic that explicitly sets
the nonce which seems to matter a lot with high concurrency in order
to avoid getting lots of "already known" errors.